Fixed enum bitpacking in basic_function.
GCC uses unsigned for enum, unless the enum includes explicit negative assignments. MSVC appears to default to signed. Bitpacking here creates errors if the enum is signed because the maximum value for 3 bits is 3, not 7.

The easy solution is to add an extra bit for signed/unsigned.

I wonder also if bitpacking is necessary?